+++ experiments/platform_test/lib/components/app.dart
@@ -8,7 +8,7 @@ import 'app.imports.dart';
 
 part 'app.g.dart';
 
-@app
+@client
 class App extends StatelessComponent with _$App {
   @override
   Iterable<Component> build(BuildContext context) sync* {