+++ packages/jaspr_builder/lib/src/client/client_registry_builder.dart
@@ -6,18 +6,19 @@ import 'package:dart_style/dart_style.dart';
 import 'package:glob/glob.dart';
 import 'package:path/path.dart' as path;
 
-class IslandsBuilder implements Builder {
-  IslandsBuilder(BuilderOptions options);
+/// Builds web entrypoints for components annotated with @app
+class ClientRegistryBuilder implements Builder {
+  ClientRegistryBuilder(BuilderOptions options);
 
   @override
   FutureOr<void> build(BuildStep buildStep) async {
     try {
-      var islandsSource = await generateIslands(buildStep);
-      if (islandsSource != null) {
-        var outputId = AssetId(buildStep.inputId.package, 'web/main.islands.dart');
+      var appsSource = await generateApps(buildStep);
+      if (appsSource != null) {
+        var outputId = AssetId(buildStep.inputId.package, 'web/main.clients.dart');
         await buildStep.writeAsString(
           outputId,
-          islandsSource,
+          appsSource,
         );
       }
     } catch (e, st) {
@@ -32,16 +33,16 @@ class IslandsBuilder implements Builder {
 
   @override
   Map<String, List<String>> get buildExtensions => const {
-        'lib/\$lib\$': ['web/main.islands.dart']
+        'lib/\$lib\$': ['web/main.clients.dart']
       };
 
   String get generationHeader => "// GENERATED FILE, DO NOT MODIFY\n"
       "// Generated with jaspr_builder\n";
 
-  Future<String?> generateIslands(BuildStep buildStep) async {
-    var islands = await buildStep.findAssets(Glob('web/**.island.dart')).toList();
+  Future<String?> generateApps(BuildStep buildStep) async {
+    var apps = await buildStep.findAssets(Glob('web/**.client.dart')).toList();
 
-    if (islands.isEmpty) {
+    if (apps.isEmpty) {
       return null;
     }
 
@@ -49,13 +50,13 @@ class IslandsBuilder implements Builder {
       $generationHeader
       
       import 'package:jaspr/browser.dart';
-      ${islands.mapIndexed((index, c) => "import '${path.url.relative(c.path, from: 'web')}' deferred as i$index;").join('\n')}
+      ${apps.mapIndexed((index, c) => "import '${path.url.relative(c.path, from: 'web')}' deferred as i$index;").join('\n')}
       
       void main() {
-        runIslandsDeferred({
-          ${islands.mapIndexed((index, c) {
+        registerClients({
+          ${apps.mapIndexed((index, c) {
       return '''
-              '${path.url.relative(path.url.withoutExtension(path.url.withoutExtension(c.path)), from: 'web')}': loadIsland(i$index.loadLibrary, 
+              '${path.url.relative(path.url.withoutExtension(path.url.withoutExtension(c.path)), from: 'web')}': loadClient(i$index.loadLibrary, 
               (p) => i$index.getComponentForParams(p),
               ),
             ''';